Update deployment finalizer to explicitely handle dependents deletion

+++ internal/controller/deployment/controller_finalize.go
@@ -20,20 +20,29 @@ package deployment
 
 import (
 	"context"
+	"errors"
 	"fmt"
+	"time"
 
+	k8sapierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/meta"
 	ctrl "sigs.k8s.io/controller-runtime"
+	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
+	"sigs.k8s.io/controller-runtime/pkg/log"
 
 	choreov1 "github.com/openchoreo/openchoreo/api/v1"
 	"github.com/openchoreo/openchoreo/internal/controller"
+	k8sintegrations "github.com/openchoreo/openchoreo/internal/controller/deployment/integrations/kubernetes"
+	"github.com/openchoreo/openchoreo/internal/labels"
 )
 
 const (
 	// DataPlaneCleanupFinalizer is the finalizer that is used to clean up the data plane resources.
 	DataPlaneCleanupFinalizer = "core.choreo.dev/data-plane-cleanup"
 )
+
+var ErrEndpointDeletionWait = errors.New("some endpoints are still finalizing, retry later")
 
 // ensureFinalizer ensures that the finalizer is added to the deployment.
 // The first return value indicates whether the finalizer was added to the deployment.
@@ -52,6 +61,7 @@ func (r *Reconciler) ensureFinalizer(ctx context.Context, deployment *choreov1.D
 
 // finalize cleans up the data plane resources associated with the deployment.
 func (r *Reconciler) finalize(ctx context.Context, old, deployment *choreov1.Deployment) (ctrl.Result, error) {
+	logger := log.FromContext(ctx).WithValues("deployment", deployment.Name)
 	if !controllerutil.ContainsFinalizer(deployment, DataPlaneCleanupFinalizer) {
 		// Nothing to do if the finalizer is not present
 		return ctrl.Result{}, nil
@@ -73,10 +83,41 @@ func (r *Reconciler) finalize(ctx context.Context, old, deployment *choreov1.Dep
 	}
 
 	resourceHandlers := r.makeExternalResourceHandlers()
+	pendingDeletion := false
+
 	for _, resourceHandler := range resourceHandlers {
+		// Skip the namespace resource as it should not be considered to handle the deletion
+		if resourceHandler.Name() == k8sintegrations.NamespaceHandlerName {
+			continue
+		}
+		// Attempt to delete the resource
 		if err := resourceHandler.Delete(ctx, deploymentCtx); err != nil {
 			return ctrl.Result{}, fmt.Errorf("failed to delete external resource %s: %w", resourceHandler.Name(), err)
 		}
+
+		// Check if the resource is still being deleted
+		exists, err := resourceHandler.GetCurrentState(ctx, deploymentCtx)
+		if err != nil {
+			return ctrl.Result{}, fmt.Errorf("failed to check existence of external resource %s: %w", resourceHandler.Name(), err)
+		}
+		if exists != nil {
+			pendingDeletion = true
+		}
+	}
+
+	// Requeue the reconcile loop if there are still resources pending deletion
+	if pendingDeletion {
+		logger.Info("endpoint deletion is still pending as the dependent resource deletion pending.. retrying..")
+		return ctrl.Result{RequeueAfter: time.Second * 5}, nil
+	}
+
+	// Clean up the endpoints associated with the deployment
+	if err := r.cleanupEndpoints(ctx, deployment); err != nil {
+		if errors.Is(err, ErrEndpointDeletionWait) {
+			// this means the endpoint deletion is still in progress. So, we need to retry later.
+			return ctrl.Result{RequeueAfter: time.Second * 5}, nil
+		}
+		return ctrl.Result{}, err
 	}
 
 	// Remove the finalizer after all the data plane resources are cleaned up
@@ -86,4 +127,57 @@ func (r *Reconciler) finalize(ctx context.Context, old, deployment *choreov1.Dep
 		}
 	}
 	return ctrl.Result{}, nil
+}
+
+func (r *Reconciler) cleanupEndpoints(ctx context.Context, deployment *choreov1.Deployment) error {
+	logger := log.FromContext(ctx).WithValues("deployment", deployment.Name)
+	logger.Info("Cleaning up the endpoints associated with the deployment")
+
+	// List all the endpoints associated with the deployment
+	endpointList := &choreov1.EndpointList{}
+	listOpts := []client.ListOption{
+		client.InNamespace(deployment.Namespace),
+		client.MatchingLabels{
+			labels.LabelKeyDeploymentName: deployment.Name,
+		},
+	}
+
+	if err := r.List(ctx, endpointList, listOpts...); err != nil {
+		if k8sapierrors.IsNotFound(err) {
+			logger.Info("No endpoints associated with the environment")
+			return nil
+		}
+		return fmt.Errorf("error listing endpoints: %w", err)
+	}
+
+	pendingDeletion := false
+
+	for _, endpoint := range endpointList.Items {
+		if err := r.Delete(ctx, &endpoint); err != nil {
+			if k8sapierrors.IsNotFound(err) {
+				// The endpoint is already deleted, no need to retry
+				continue
+			}
+			return fmt.Errorf("error deleting endpoint %s: %w", endpoint.Name, err)
+		}
+
+		// Get the resource back to check if the resource still exists
+		if err := r.Get(ctx, client.ObjectKeyFromObject(&endpoint), &endpoint); err != nil {
+			if k8sapierrors.IsNotFound(err) {
+				// The endpoint is already deleted, no need to retry
+				continue
+			}
+			return fmt.Errorf("error getting endpoint %s: %w", endpoint.Name, err)
+		}
+
+		// marking blocked as true as the endpoint has still not deleted.
+		pendingDeletion = true
+	}
+
+	// If at least one endpoint is blocked, signal that we need to retry later
+	if pendingDeletion {
+		return ErrEndpointDeletionWait
+	}
+
+	return nil
 }
